requirements.txt
adal==1.2.3
arrow==0.15.2
astroid==2.2.5
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==2.9.2
decorator==4.4.0
Django==2.1.15
django-model-utils==4.0.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.0
idna==2.9
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.1
parso==0.5.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==2.0.9
pycparser==2.20
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyodbc==4.0.32
python-dateutil==2.8.0
python-memcached==1.59
pytz==2020.1
requests==2.23.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
timeago==1.0.13
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.9
wcwidth==0.1.7
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.9
